started to add jwt token

+++ SportClubs/Entities/Club.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string? Description { get; set; }
-        public double TeacherId { get; set; }
+        public int TeacherId { get; set; }
         public Teacher Teacher { get; set; }
         public Schedule Schedule { get; set; }
         public List<StudentClub> StudentClubs { get; set; }
